WIP: The openocd-0.11.0 release
No plan to merge this, but this initial part is the most tedious
and boring task to bump a new release and seams being also one of
the blocking points to proceed with a new release.
So, here is a preliminarily version, mainly aimed at getting your
review. There are today 1266 commits merged after v0.10.0, and I'm
pretty sure I did not documented properly some of them.
Feel free to address the mistakes in gerrit.
I will try to update and rebase this change while more commits get
merged, until there is consensus on tagging v0.11.0.

By the way, v0.11.0 or v1.0.0 ?

Change-Id: I111fec1304482f5c0f9d6ee988be8a2ea3de3981
Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>