targets: unconditionally invalidate registers cache on reset_assert
In case of deferred target examination, the flag for target examined is
cleared in jim_target_reset() _before_ calling assert_reset(). The
register cache is then never invalidated, even if the target was
manually examined before the reset.

Remove the test and unconditionally invalidate the registers cache
during assert_reset(), as implemented for the other target types.

Change-Id: Ifd0a735fe6afd332829e30c32f1016735dbe2d20
Signed-off-by: Mikhail Rasputin <mikhail.godlike.rasputin@yandex.ru>