target/armv7m: add a configurable TPIU base address
The base-address of an on-core TPIU is usually 0xE0040000.
This may change when the TPIU is shared between multiple cores.

To cover the second case, the TPIU registers are now accessed
by (base-address + TPIU_REGISTER_OFFSET).

The "tpiu" command has a new optional parameter, "baseaddr <base address>".
When left unspecified, it has a default value of 0xE0040000.

Change-Id: I7c4041636e0df7623e1758905c5fbb9db294c422
Signed-off-by: Adrian Negreanu <adrian.negreanu@nxp.com>