topic: Added support to debug ARM AARCH32 USER mode on AARCH64
The code to debug (set breakpoints, inspect memory, etc.) for
the ARM AARCH32 USER mode was missing.
E.g., Debugging 32-bit code that runs in user mode (EL0) on an
AARCH64 ARM processor (e.g., on a Raspberry Pi 4) was not possible.

This patch adds support to do this.
In aarch64.c switching to SVC mode was added when in USR mode in
order to execute the privileged instructions (e.g., mmu modify).
In armv8_cache.c the flushing / invalidating of the caches was
added for 32-bit mode. I.e. switch to SVC mode if in USR mode,
execute DCCIMVAC / ICIMVAU (AARCH32) instead of DC CIVAC / IC IVAU
(AARCH64).

This now allows to set software breakpoints in AARCH32 USER mode.
This also allows stepping over functions or syscalls in AARCH32 USER
mode with GDB. This also allows to inspect the memory in AARCH32
USER mode from GDB.

Change-Id: I9c20e674837ce10ee605ef7a5f1563bd84e95189
Signed-off-by: Rick Wertenbroek <rick.wertenbroek@heig-vd.ch>