board: add Skytone Alpha 400
This adds the configuration for the MIPS subnotebook board that is known
by many names -- Skytone Alpha 400 seems to be the OEM name. The unit
I've tested this on was sold as "Elonex Onet+".

Enough plafrom initialization is provided to make UART0 (routed to
connector in battery compartment) and DRAM and access to NAND flash chip
accesible.

Change-Id: If21177ed7839e266bfd8f1fc397d4c04a3174dd1
Signed-off-by: Lubomir Rintel <lkundrak@v3.sk>
Signed-off-by: Jiri Kastner <cz172638@gmail.com>