stm32l5xx: support flash programming while TrustZone is enabled
STM32L5 flash memory is aliased to 0x0C000000, this address mapping
is used for secure applications. (0x08000000 for non-secure)

this change allows the programming of secure and non-secure flash
when TZEN=1 and RDP level is 0

When TZEN=1 and RDP level is 0.5, the work_area could be protected
with secure watermarks, unfortunately there is no automatic way to
detect the non-protected RAM to use as work_area.

Change-Id: I89d1f1b5d493cf01a142ca4dbfef5a3731cab96e
Signed-off-by: Tarek BOCHKATI <tarek.bouchkati@st.com>