DO NOT MERGE! checkpatch: remove FSF check. DO NOT MERGE!
DO NOT MERGE!!! DO NOT MERGE!!! DO NOT MERGE!!! DO NOT MERGE!!!

For the only purpose of allowing Jenkins to check the following
patches in this series, let's remove from checkpatch the test
for the Free Software Foundation address. This test fails on two
of the patches in the series because they:
- move the GPL-v2 license in another path,
- add the GFDL-v1.2 license,
thus they MUST contain the FSF address.

This patch is NOT TO BE MERGED. Should be 'abandoned' when the
remaining patches of this series would be be merged.

DO NOT MERGE!!! DO NOT MERGE!!! DO NOT MERGE!!! DO NOT MERGE!!!

Change-Id: I9a4a2ba8fa7a1ff86876cf720c9f0c55f7bb060d
Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>