configure.ac: drop macro 'AC_PROG_CC_C99'
The macro AC_PROG_CC_C99 has been obsoleted by autoconf 2.70 and
its test is now included in the macro AC_PROG_CC. The latter macro
is already present in configure.ac.
For autoconf 2.69 and earlier, the macro has to be specified if
needed.

The macro AC_PROG_CC_C99 checks if the C compiler is not in C99
mode by default, then searches for a command line flag able to put
the compiler in C99 mode.
The flags tested are:
'-std=gnu99', '-std=c99', '-c99', '-AC99', '-D_STDC_C99=' and
'-qlanglvl=extc99'.

From gcc 5.1.0 (2015-04-22) the default when no command line flag
is given is to use C11. Later versions of gcc use C99 as default.

As old as gcc 5.1.0, also clang 3.6.0 (2015-02-27) uses C11 as
default. Later versions of clang use C17 as default.

So, starting from gcc 5.1.0 and clang at least 3.6.0, the macro
AC_PROG_CC_C99 is not required anymore.

For users on compilers other than gcc, the autoconf macro could
still be useful, but today gcc and clang are the two major
compilers used for OpenOCD on all platforms. The risk to break the
build removing the macro is really low.

Remove the macro AC_PROG_CC_C99 to avoid the autoconf warning for
obsoleted macro.

Change-Id: I5e8437f5a826fb63be6d07bcb5bb824f94683020
Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>