jlink: fix device discovery when network is off
If user specifies a serial number for the jlink device, openocd
extends the search to network jlink devices too, without checking
if the host has a valid and functional network connection. If the
network is not functional, libjaylink returns error. This error
invalidates the discovery on USB, even if it was successful.

Split the discovery on USB and network, return error only if both
cases fail, while ignore the error if at least one case succeed.

Change-Id: I537a44ba9551b6e43872a60b94395acb749ff6ba
Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>
Fixes: https://sourceforge.net/p/openocd/tickets/294/
Reported-by: <Bohdan Tymkiv <bohdan200@users.sourceforge.net>