Makefile.am: fix for old GNU make 3.82
Commit 4fc0f3530c51 ("Makefile.am: fix non-POSIX warning from
automake") uses the BSD make extension '!=' to generate the list
of files NEWS-xx, but it's not POSIX (yet), it's not backward
compatible with GNU make 3.82 (used in CentOS 7, EOL scheduled for
2024) and depending on automake evolution it could trigger again
the warning 'apparently' just fixed.

Remove completely the shell invocation to build the variable and
simply use a filename wildcard.

The make's variable EXTRA_DIST will _NOT_ contain the expanded
filename, but simply the string 'NEWS-*'. It will be used either
as a pre-requisite of a make target and in the recipe of the same
target. For the recipe there is no problem, because it's a shell
code fragment where filename wildcard will be properly expanded.
For the prerequisite it 'seems' fine too; POSIX is not clear on
this point, but so far most 'make' implementations accept filename
expansion in the prerequisites and GNU make reports it in the user
manual (chapter 4.3 'Using Wildcard Characters in File Names').
See also: https://unix.stackexchange.com/questions/513644/

Change-Id: I83064e08e15d7182c5d99962b883f7e1a960dbee
Signed-off-by: Andreas Fritiofson <andreas.fritiofson@gmail.com>
Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>
Reported-by: Jan Matyáš <jmatyas@codasip.com>