gdb_server: fix detection of target halt after gdb 'continue'
OpenOCD is based on a single main loop that schedules all the
activities.
At the execution of a TCL command, the timestamp is checked to
eventually trigger the polling of the targets. This means that by
executing a TCL command the polling can be triggered and detect a
change of target's state.
This can be a problem if a TCL command has to be executed during a
code sequence that is already changing the target's state (e.g.
halt, resume or single-step) and the undesired poll can cause a
lost of sync between the code sequence and openocd target's state.

The gdb command 'continue' causes gdb_handle_vcont_packet() to run
target_resume(), that in turn runs the target event 'resumed'
after the successful restart of the target.
If an handler is associated to the event above, the execution of
the TCL commands in the handler can trigger the polling.
The target is just restarted, but could have already hit another
breakpoint so the new state, that should be 'running', could be
again 'halted'!
This early return in 'halted' state during the resume process
prevents gdb_server to inform gdb about the new halt, and gdb
keeps waiting for target to halt.

This can be verified by running on the target a firmware with a
loop, run openocd with a dummy 'resumed' event, and let gdb to set
a breakpoint in the loop. A 'continue' command will cause the
target to halt again by hitting the breakpoint at the next loop
iteration, but gdb will loose it and will not return the prompt.

	openocd -f board/st_nucleo_f4.cfg -c \
		'stm32f4x.cpu configure -event resumed {echo hello}'
	arm-none-eabi-gdb -ex 'target remote :3333' -ex 'b *$pc' -ex c

Disable the polling while executing target_resume().

Change-Id: I3be830a8e7c2ef6278617cb4547a4d676b0ddeb5
Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>
Reported-by: Габитов Александр Фаритович <gabitov@planarchel.ru>
Fixes: d301d8b42f0b ("gdb_server: add support for vCont")