cortex_m: clear halt debug enable bit in cortex_m_deinit_target()
Before exiting cortex_m_deinit_target(), write to DHCSR to clear
the DHCSR control bits. The focus is on clearing DHCSR.DEBUGEN,
which disables the halt debug mode on the CPU.

This function is needed in some use cases, where openocd is being
used to flash and reset a target device, which shall then execute
a program directly, that is, without doing power-on reset  (which
would, otherwise, clear the DHCSR.DEBUGEN bit). The DEBUGEN flag
must be cleared, if the program requires to actively handle debug
events in run-time. An example of this case are debug events that
are generated by DWT, and handled via the DebugMonitor exception:
with DEBUGEN set to 1, such events will halt the CPU, instead of
triggering the DebugMonitor. Note that DAP is the only way the
DEBUGEN can be cleared without power-on reset; this bit cannot be
written by software.

Change-Id: I90c5f7ca085c9e1a211b0a4dc0e0ff6f0a54889d
Signed-off-by: Ioannis Glaropoulos <Ioannis.Glaropoulos@nordicsemi.no>