cortex_m: clear halt debug enable bit in cortex_m_deinit_target
Before exiting cortex_m_deinit_target(), write to DHCSR to clear
the DHCSR control bits (primarily DHCSR.DEBUGEN), to disable the
halt debug mode. This is required for the target to be in normal
(i.e. not debug) mode after the session shuts down. The reason for
this is that DHCSR.DEBUGEN can only be cleared by DAP (not software)
or after a power-on reset.

Change-Id: I90c5f7ca085c9e1a211b0a4dc0e0ff6f0a54889d
Signed-off-by: Ioannis Glaropoulos <Ioannis.Glaropoulos@nordicsemi.no>