target: add event 'shutdown'
OpenOCD does not modify anything on the target when it shutdowns.
If the target was halted or running, it will keep the same.
Any configuration for target clock, cpu debug enable, power
management and so on will be left in the current state.

This is not always ideal, but any action at OpenOCD shutdown is
strongly related to the use case, and cannot be standardized.

Add a new target event 'shutdown' that would be called before
OpenOCD quits. It can contain any action that the user considers
relevant before OpenOCD quits.

Change-Id: Ibc3804f232817a454835e151c2a5b62b3f7b52fc
Signed-off-by: Antonio Borneo <borneo.antonio@gmail.com>