target/once: Initial support for On Chip Emulator (Power arch, e200 core)
Basic support for debug control of the Power Architecture e200 core.
All of the SPC56x/RPC56x devices from ST (and possibly MPC56x from NXP) should
be compatible with this implementation, based on ST Application Note AN4035.

The following functionality is provided:
 - Read/Write CPU registers
 - Read/Write arbitrary memory location (32-bit)
 - Execute single instruction
 - Load binary file to memory location
 - Execute code at specific memory location, until a breakpoint is reached

Tested on ST SPC560D.

binary.tcl is added as a dependency. It is taken from jimtcl. The full original
license is added on the header, plus an SPDX meta tag, with GPL2 as extra choice

Change-Id: Icaa8a5d38923f5768e64f00c62058fafe93ab6ed
Signed-off-by: Luis de Arquer <luis.dearquer@inertim.com>