COVER LETTER: Initial support for ST On Chip Emulator (OnCE)
This patchset provides basic support for debug control of the
Power Architecture e200 core.

All of the SPC56x/RPC56x devices from ST (and possibly MPC56x
from NXP) should be compatible with this implementation, based on
ST Application Note AN4035.

The following functionality is provided:
 - Read/Write CPU registers
 - Read/Write arbitrary memory location (32-bit)
 - Execute single instruction
 - Load binary file to memory location
 - Execute code at given memory location, until breakpoint is reached

Tested on ST SPC560D.

Even though OnCE can be implemented mostly by regular IRSCAN and
DRSCAN instructions, there are some nasty particularities which
required low level changes, and which may not be to everyone's taste
in its current form.

1) OnCE Status Register (OSR): This register is read in a very
special way: it is shifted out serially during an IRSCAN. To do so,
Commit 1 changes the IRSCAN tcl implementation so that it can
return a value. Maybe this could be implemented by providing an
extra option to IRSCAN.

2) If DR-PAUSE is entered before DR-UPDATE, then OnCE TAP gets
disabled. Unfortunately, this is the default behaviour in OpenOCD.
Commit 2 proposes a way to adapt for this (currently only implemented
on the FTDI driver), with the naive hope of not breaking anything.
As with the change on IRSCAN, maybe this could be enabled by an extra
option to the TCL command.

Commit 3 creates a new TCL file, once.cfg. This is the core of the
patchset, and feedback is welcome on how to shape it.

Commit 4 provides a board target file, for the ST SPC560D discovery
board, providing an example on how to initialize the OnCE interface.

The TCL code is not meant to be exhaustive yet (e.g. error handling
and documentation are still missing). I am looking for feedback at
this stage, rather than providing a polished implementation which may
not fit OpenOCD practices (for example, maybe it makes more sense to
code once.cfg in C)

Change-Id: Id821afb1a57276aa911331ef0d75e956d4c305a4
Signed-off-by: Luis de Arquer <luis.dearquer@inertim.com>