Cleanup of config/includes.
Remove a use of AH_BOTTOM from configure.ac. This macro was used to
include a few files into every file in the project. That causes problems
when I want to use the gnulib library (which comes with its own
Makefile, and does not have the same include path as the top-level
Makefile), and the simplest solution I could think of was to explicitly
include the files where they're required.

So this change touches a lot of files, but is actually really simple.
While this in preparation for using gnulib, this change does not affect
functionality at all.

Change-Id: I52c70bf15eb2edc1dd10e0fde23b2bcd4caec000
Signed-off-by: Tim Newsome <tim@sifive.com>