WIP: support EFM32 series 2 devices
Preliminary support for, at least, the EFR32xG22 Series 2 devices from
SiLabs.  These are Cortex-M33, and have a very different "MSC"
block.  At this stage, debugging is supported, and most of what should
be necessary for flash operations is supported, but I run into problems
with the core not halting before it even tries flash operations :(

The offsets of MSC registers have changed, so the existing loader cannot
work, and has been set to fallback to word by word for now.

Lock bits reading is implemented for normal pages, but not for any of
the special pages, and lock bits writing is completely unsupported.

Printing out nice part names is not finished, unsure how to chain the
descriptors.

Change-Id: If9fd09b002e6c531a76963054847580b65180371
Signed-off-by: Karl Palsson <karlp@tweak.net.au>