target/arm_adi_v5: Fixes for Renesas RZ/G2L
This fix to mem_ap functions is required to support RZ/G2L. The
problem is related to the back to back CSW write and TAR, that
causes a WAIT response from DAP. Folling commands, even if
repeated later in the correct order cause problems to the DAP
itself. Solved by ensuring that as soon as a WAIT is received
the commands are repeated and others are not sent.

Change-Id: Ia0dcbdd8b975083e2462976ccdca26875a1add31
Signed-off-by: micbis <michele.bisogno.ct@renesas.com>