target/arm_adi_v5: Fixes for Renesas RZ/G2L
This fix to mem_ap_init is required to support RZ/G2L. The first
problem is related to the back to back CSW write and successive
reads to CSW and CFG, solved by splitting the accesses. The second
is related to the CSW MEM-AP access size. 8-bit is not required by
Arm ADI v5 architecture and it causes troubles when examining the
Cortex-M33. Fixed by using 32-bit access size, whose support is
mandatory.

Change-Id: Ia0dcbdd8b975083e2462976ccdca26875a1add31
Signed-off-by: micbis <michele.bisogno.ct@renesas.com>