target/aarch64: Write Memory functions fix for self-modifying code
Writing memory function shall invalidate / flush caches. This is required
for self-modifying code. The I/D cache invalidate code shall not check
for the caches to be enabled. In AArch64 the SCTLR I and C bits do NOT
disable the caches, they prevent code/data allocation. So if some stale
code/data is present in the cache the CPU may (incorrectly) use it.

Change-Id: I3bcd21770377be60a3788d66cd2a118aef1df403
Signed-off-by: micbis <michele.bisogno.ct@renesas.com>