tcl/target: Select default boot core
On SMP Renesas R-Car Gen2/Gen3 systems, select the boot core as the
default target the 'targets' command. This way, the user can start
debugging code running on the boot core without having to switch to
the boot core by explicitly invoking 'targets' command first, since
it is likely the debugged code will run on the boot core. Note that
most of the code is already in place, it was just not used, so this
is more of a fix to make the original intention work.

Change-Id: I727808adce617c1d9ebd6ffa34f60f5882cdae60
Signed-off-by: Marek Vasut <marek.vasut@gmail.com>