arm_adi_v5: Support parsing nested ROM tables
ADI v5.1 and v6.0 permit the definition of CoreSight components (class 9
ROM entries) with a device type value of 0 which is to be scanned as a
nested ROM table. Class 9 (CoreSight) components can only have up to
512 ROM table entries and the present indicator is with the value 0b11.

dap_rom_display() is refactored a bit such that we always end up with
attempting to parse the ROM contents using the appropriate upper limit
for class 1 and 9 ROM types.

Change-Id: I4ba497b3807f1f11f06186eb6e61959ea3540c59
Signed-off-by: Florian Fainelli <f.fainelli@gmail.com>