arm_adi_v5: Recurse looking up a CS if we have a nested ROM
Similar to the previous commit which fixed the "dap info" displaying of
nested ROMs, when we encounter a device type of 0 in the Class 9 ROMs,
we need to continue and recurse until we can look up the desired DAP CS
with the expected type (0x15).

Towards that goal, move the reading of the device type a bit earlier so
we can key both the class type (1 or 9) and the device type to trigger a
nested ROM parsing if needed.

This fixes the inability to debug, e.g.: Cortex-A76 CPUs since we were
not able to look up the CPU CS before.

Change-Id: I8efacf8b776d276e39cdbfaa64d66cf8d02aca05
Signed-off-by: Florian Fainelli <f.fainelli@gmail.com>